pyproject.toml
[tool.poetry]
name = "zuzubliks-bot"
version = "0.1.0"
description = "Телеграм для вычесления средней цены"
authors = ["https://t.me/Staks_sor"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
python-telegram-bot = "^21.0"
pandas = "^2.2.0"
openpyxl = "^3.1.0"
numpy = "^1.26.0"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"